+++ pykakasi/jisyo.py
@@ -3,7 +3,7 @@
 #
 # Copyright 2011-2018 Hiroshi Miura <miurahr@linux.com>
 from pkg_resources import resource_filename
-from six.moves.cPickle import load
+from six.moves import cPickle
 
 class jisyo (object):
     _dict = None
@@ -11,7 +11,7 @@ class jisyo (object):
 
     def __init__(self, pklname):
         with  open(resource_filename(__name__, pklname), 'rb') as dict_pkl:
-            (self._dict, self._len) = load(dict_pkl)
+            (self._dict, self._len) = cPickle.load(dict_pkl)
 
     def haskey(self, key):
         return key in self._dict